Obsługa błedu AritmethicExecption...

+++ src/main/java/Calc/Calculator.java
@@ -2,24 +2,22 @@ package Calc;
 
 public class Calculator {
 
-    public double add(double a, double b) {
+    public int add(int a, int b) {
         return a+b;
     }
 
-    public double subtract(double a, double b) {
+    public int subtract(int a, int b) {
         return a-b;
     }
 
-    public double multiply (double a, double b) {
+    public int multiply (int a, int b) {
         return a*b;
     }
 
-    public double divide (double a, double b) {
-        return a/b;
-    }
+    public int divide (int a, int b) {return a / b;}
 
-    public double MathPow(double a, double b) {
-        return Math.pow(a,b);
+    public int MathPow(int a, int b) {
+        return (int) Math.pow(a,b);
     }
